non Gmail/Gmail Vacancy emails
Using vacancy as it allows us to find jobs for Selligence. Currently hard coded job title to make sure it is definitely the selligence job we are going into. Otherwise the UI loads out of order and you sometimes end up in the 'first' unsearched job instead of the one that we want. Can't get around this without hard waits as there is nothing to wait against.