Dropping sequencing out temporarily
Whilst I am getting to the bottom of the failures I am dropping this test out. Doesn't seem to be a bug but if you create a schedule and try to create a step in a sequence to fast then the schedules drop down is empty and obv therefore can't be selected. Seems a typical slow Staging issue